% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_fit.R
\name{mp_fit}
\alias{mp_fit}
\title{fit median polish to a long form table}
\usage{
mp_fit(long_df, measure, row_name, col_name)
}
\arguments{
\item{long_df}{a long data frame with a numeric 'measure', and 'row_name' and 'col_name' classifiers}

\item{measure}{a numeric variable in long_df}

\item{row_name}{a variable in long_df}

\item{col_name}{a variable in long_df}
}
\value{
lp
}
\description{
fit median polish to a long form table
}
\examples{
mp_fit(long_df, measure, row_name, col_name)
}
