% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_plot.R
\name{mp_plot}
\alias{mp_plot}
\title{plot expanded medpol output from mp_fit}
\usage{
mp_plot(expanded_lp_list)
}
\arguments{
\item{expanded_lp_list}{augmented output from medpolish returned from the mp_fit function}
}
\value{
p

The dataframe with new mean and sum columns
}
\description{
plot expanded medpol output from mp_fit
}
\examples{
mp_plot(expanded_lp_list)
}
