% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_rotate_plot.R
\name{mp_rotate_plot}
\alias{mp_rotate_plot}
\title{mp_rotate_plot}
\usage{
mp_rotate_plot(expanded_lp_list)
}
\arguments{
\item{expanded_lp_list}{a ggplot glob produced by the mp_plot function}
}
\value{
vp
}
\description{
mp_rotate_plot
}
\examples{
mp_rotate_plot(expanded_lp_list)
}
